feat(settings): Re-organize settings
Kind of a big one:
 - Fixes a few missing settings from 05-private.py
 - Reorganizes settings into directories
 - Removes a few settings that are only used in a few places
