+++ cl/settings/project/maintenance_mode.py
@@ -1,0 +1,8 @@
+MAINTENANCE_MODE_ENABLED = False
+MAINTENANCE_MODE_ALLOW_STAFF = True
+MAINTENANCE_MODE_ALLOWED_IPS = []
+MAINTENANCE_MODE = {
+    "enabled": MAINTENANCE_MODE_ENABLED,
+    "allow_staff": MAINTENANCE_MODE_ALLOW_STAFF,
+    "allowed_ips": MAINTENANCE_MODE_ALLOWED_IPS,
+}
